requirements.txt
Django==4.1
django-crispy-forms==1.14.0
django-csp==3.7
django-environ==0.9.0
django-sri==0.4.0
django-heroku
gunicorn==20.1.0
Pillow==9.2.0
whitenoise
psycopg2==2.9.3